+++ database.js
@@ -42,7 +42,7 @@ export class Database {
 
     /**
      * Get the database name.
-     * @return {String} Then name of the database.
+     * @return {String} The name of the database.
      */
     get name() {
         // Return the name value
@@ -88,9 +88,12 @@ export class Database {
             }
             
             // Handle on error event
-            openDbRequest.onerror = () => {
+            openDbRequest.onerror = (event) => {
                 // Clear interface value
                 this._iDbDatabase = undefined;
+
+                // If connection was blocked then do nothing
+                if (event.target._blocked === true) return;
 
                 // If there is an upgrade error then reject with that instead of the error
                 if (upgradedError)  { reject(upgradedError); return; }
@@ -100,7 +103,14 @@ export class Database {
             };
 
             // Handle on success event
-            openDbRequest.onsuccess = () => {
+            openDbRequest.onsuccess = (event) => {
+                // If connection was blocked
+                if (event.target._blocked === true) {
+                    // Close the database and stop
+                    openDbRequest.result.close();
+                    return;
+                }
+
                 // Set the database interface
                 this._iDbDatabase = openDbRequest.result;
 
@@ -122,6 +132,17 @@ export class Database {
 
             // Handle on upgrade needed event
             openDbRequest.onupgradeneeded = (event) => {
+                // If connection was blocked
+                if (event.target._blocked === true) {
+                    // We have already rejected this database connect, and this "openDbRequest"
+                    // object is kept around until the database connection blocking it was closed, so
+                    // it may have been hanging around for some time.
+
+                    // Close the database and stop
+                    openDbRequest.result.close();
+                    return;
+                }
+
                 // Set the database interface
                 this._iDbDatabase = event.target.result;
 
@@ -159,8 +180,11 @@ export class Database {
 
             // Handle on blocked event
             openDbRequest.onblocked = () => {
-                // Reject the promise with the error
-                reject(openDbRequest.error);
+                // Set blocked flag
+                openDbRequest._blocked = true;
+                
+                // Reject the promise with "blocked" error
+                reject(new Error('Blocked'));
             }
         });